refactor: move connection tab logic back to own comp
