+++ templates/vue/src/components/tyde/activities/CircleOfSupport/CommunityView/ConnectionsTab.vue
@@ -1,37 +1,216 @@
 <template>
-  <cos-popup :show="show">
+  <cos-popup :show="isOpen">
     <template #toggle>
       <cos-popup-button
         style="left: 2rem"
         aria-label="Connections"
-        @click="$emit('toggle')"
+        @click="isOpen = !isOpen"
       >
-        <tapestry-icon v-if="show" icon="chevron-down" />
+        <tapestry-icon v-if="isOpen" icon="chevron-down" />
         <span v-else>😊</span>
       </cos-popup-button>
     </template>
     <template #content>
-      <slot></slot>
+      <div
+        v-if="state === states.Add || state === states.Edit"
+        class="content-wrapper"
+      >
+        <b-overlay class="form" :show="isSubmitting">
+          <add-connection-form
+            v-model="connection"
+            :communities="communities"
+            @back="back"
+            @submit="handleSubmit"
+            @add-community="$emit('add-community', $event)"
+          />
+        </b-overlay>
+      </div>
+      <connections-list
+        v-else
+        :connections="connections"
+        :communities="communities"
+        @add-connection="openConnectionForm"
+        @edit-connection="editConnection"
+      />
     </template>
   </cos-popup>
 </template>
 
 <script>
 import TapestryIcon from "@/components/common/TapestryIcon"
+import client from "@/services/TapestryAPI"
+
 import CosPopup from "./CosPopup"
 import CosPopupButton from "./CosPopupButton"
+import AddConnectionForm from "./AddConnectionForm"
+import ConnectionsList from "./ConnectionsList"
+
+const States = {
+  Home: 0,
+  Add: 1,
+  Edit: 2,
+}
 
 export default {
   components: {
     CosPopup,
     CosPopupButton,
+    AddConnectionForm,
+    ConnectionsList,
     TapestryIcon,
   },
   props: {
-    show: {
-      type: Boolean,
+    connections: {
+      type: Object,
       required: true,
+    },
+    communities: {
+      type: Object,
+      required: true,
+    },
+  },
+  data() {
+    return {
+      isOpen: false,
+      state: States.Home,
+      isSubmitting: false,
+      connection: {
+        id: "",
+        name: "",
+        avatar: "😊",
+        communities: [],
+      },
+    }
+  },
+  computed: {
+    states() {
+      return States
+    },
+  },
+  methods: {
+    back() {
+      this.state = States.Home
+      this.$emit("back")
+    },
+    open() {
+      this.isOpen = true
+    },
+    close() {
+      this.isOpen = false
+    },
+    openConnectionForm() {
+      this.resetConnection()
+      this.state = States.Add
+    },
+    editConnection(connection) {
+      this.open()
+      this.connection = {
+        ...connection,
+        communities: connection.communities.map(community => community.id),
+      }
+      this.state = States.Edit
+    },
+    resetConnection() {
+      this.connection = {
+        id: "",
+        name: "",
+        avatar: "😊",
+        communities: [],
+      }
+    },
+    async handleSubmit() {
+      this.isSubmitting = true
+
+      switch (this.state) {
+        case States.Add:
+          await this.addNewConnection()
+          break
+        case States.Edit:
+          await this.updateConnection()
+          break
+        default:
+          break
+      }
+
+      this.isSubmitting = false
+      this.resetConnection()
+      this.state = States.Home
+    },
+    async addNewConnection() {
+      const connection = await client.cos.addConnection({
+        name: this.connection.name,
+        avatar: this.connection.avatar,
+      })
+
+      if (this.connection.communities.length) {
+        /**
+         * Add connection to community one at a time to avoid race condition where
+         * only the last community is kept.
+         */
+        for (const communityId of this.connection.communities) {
+          await client.cos.addConnectionToCommunity(communityId, connection.id)
+        }
+      }
+      this.$emit("add-connection", {
+        ...connection,
+        communities: this.connection.communities,
+      })
+    },
+    async updateConnection() {
+      const currentCommunities = this.getCommunities(this.connection.id)
+      await client.cos.updateConnection(this.connection.id, { ...this.connection })
+
+      const { additions, deletions } = this.getDifferences(
+        currentCommunities.map(community => community.id),
+        this.connection.communities
+      )
+
+      for (const addition of additions) {
+        await client.cos.addConnectionToCommunity(addition, this.connection.id)
+      }
+
+      for (const deletion of deletions) {
+        await client.cos.removeConnectionFromCommunity(deletion, this.connection.id)
+      }
+
+      this.$emit("edit-connection", {
+        ...this.connection,
+        additions,
+        deletions,
+      })
+    },
+    getCommunities(connectionId) {
+      return Object.values(this.communities).filter(community =>
+        community.connections.includes(connectionId)
+      )
+    },
+    getDifferences(original, newVersion) {
+      const additions = newVersion.filter(item => !original.includes(item))
+      const deletions = original.filter(item => !newVersion.includes(item))
+      return { additions, deletions }
     },
   },
 }
 </script>
+
+<style lang="scss" scoped>
+ul {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
+
+.form {
+  width: 100%;
+  height: 100%;
+}
+
+.content-wrapper {
+  background: white;
+  position: relative;
+  z-index: 10;
+  height: 100%;
+  border-top: 1px solid var(--cos-color-tertiary);
+  flex-grow: 1;
+}
+</style>
