+++ templates/vue/src/components/tyde/activities/CircleOfSupport/CommunityView/index.vue
@@ -7,32 +7,14 @@
       @edit-community="editCommunity"
     />
     <connections-tab
+      ref="connections"
       class="tab"
-      :show="isConnectionTabOpen"
-      @toggle="toggleConnectionTab"
-    >
-      <div
-        v-if="state === states.AddConnection || state === states.EditConnection"
-        class="content-wrapper"
-      >
-        <b-overlay class="form" :show="isSubmitting">
-          <add-connection-form
-            v-model="connection"
-            :communities="communities"
-            @back="state = lastState"
-            @submit="handleSubmit"
-            @add-community="$emit('add-community', $event)"
-          />
-        </b-overlay>
-      </div>
-      <connections-list
-        v-else
-        :connections="connections"
-        :communities="communities"
-        @add-connection="openConnectionForm"
-        @edit-connection="editConnection"
-      />
-    </connections-tab>
+      :connections="connections"
+      :communities="communities"
+      @back="handleBack"
+      @add-connection="$emit('add-connection', $event)"
+      @edit-connection="handlEditConnection"
+    />
     <add-community-tab
       v-model="community"
       :show="isCommunityTabOpen"
@@ -45,30 +27,22 @@
 </template>
 
 <script>
-import client from "@/services/TapestryAPI"
-
 import AddCommunityTab from "./AddCommunityTab"
 import ConnectionsTab from "./ConnectionsTab"
 import CommunitiesList from "./CommunitiesList"
-import AddConnectionForm from "./AddConnectionForm"
-import ConnectionsList from "./ConnectionsList"
 import { MAX_COMMUNITIES } from "../cos.config"
 
 const States = {
   Home: 0,
-  List: 1,
-  AddConnection: 2,
+  AddCommunity: 1,
+  EditCommunity: 2,
   EditConnection: 3,
-  AddCommunity: 4,
-  EditCommunity: 5,
 }
 
 export default {
   components: {
     AddCommunityTab,
-    AddConnectionForm,
     CommunitiesList,
-    ConnectionsList,
     ConnectionsTab,
   },
   props: {
@@ -85,13 +59,6 @@ export default {
     return {
       state: States.Home,
       lastState: States.Home,
-      isSubmitting: false,
-      connection: {
-        id: "",
-        name: "",
-        avatar: "😊",
-        communities: [],
-      },
       community: {
         name: "",
         icon: "👨‍👩‍👦",
@@ -102,11 +69,6 @@ export default {
   computed: {
     states() {
       return States
-    },
-    isConnectionTabOpen() {
-      return [States.List, States.AddConnection, States.EditConnection].includes(
-        this.state
-      )
     },
     isCommunityTabOpen() {
       return [States.EditCommunity, States.AddCommunity].includes(this.state)
@@ -121,13 +83,6 @@ export default {
     },
   },
   methods: {
-    toggleConnectionTab() {
-      if (this.isConnectionTabOpen) {
-        this.state = States.Home
-      } else {
-        this.state = States.List
-      }
-    },
     toggleCommunityTab() {
       if (this.isCommunityTabOpen) {
         this.state = States.Home
@@ -136,20 +91,9 @@ export default {
         this.state = States.AddCommunity
       }
     },
-    openConnectionForm() {
-      this.resetConnection()
-      this.state = States.AddConnection
-    },
     editCommunity(community) {
       this.community = community
       this.state = States.EditCommunity
-    },
-    editConnection(connection) {
-      this.connection = {
-        ...connection,
-        communities: connection.communities.map(community => community.id),
-      }
-      this.state = States.EditConnection
     },
     resetCommunity() {
       this.community = {
@@ -158,84 +102,19 @@ export default {
         color: "",
       }
     },
-    resetConnection() {
-      this.connection = {
-        id: "",
-        name: "",
-        avatar: "😊",
-        communities: [],
+    editConnection(connection) {
+      this.state = States.EditConnection
+      this.$refs.connections.editConnection(connection)
+    },
+    handleBack() {
+      if (this.state === States.EditConnection) {
+        this.state = States.Home
+        this.$refs.connections.close()
       }
     },
-    async handleSubmit() {
-      this.isSubmitting = true
-
-      switch (this.state) {
-        case States.AddConnection:
-          await this.addNewConnection()
-          break
-        case States.EditConnection:
-          await this.updateConnection()
-          break
-        default:
-          break
-      }
-
-      this.isSubmitting = false
-      this.resetConnection()
-      this.state = this.lastState
-    },
-    async addNewConnection() {
-      const connection = await client.cos.addConnection({
-        name: this.connection.name,
-        avatar: this.connection.avatar,
-      })
-
-      if (this.connection.communities.length) {
-        /**
-         * Add connection to community one at a time to avoid race condition where
-         * only the last community is kept.
-         */
-        for (const communityId of this.connection.communities) {
-          await client.cos.addConnectionToCommunity(communityId, connection.id)
-        }
-      }
-      this.$emit("add-connection", {
-        ...connection,
-        communities: this.connection.communities,
-      })
-    },
-    async updateConnection() {
-      const currentCommunities = this.getCommunities(this.connection.id)
-      await client.cos.updateConnection(this.connection.id, { ...this.connection })
-
-      const { additions, deletions } = this.getDifferences(
-        currentCommunities.map(community => community.id),
-        this.connection.communities
-      )
-
-      for (const addition of additions) {
-        await client.cos.addConnectionToCommunity(addition, this.connection.id)
-      }
-
-      for (const deletion of deletions) {
-        await client.cos.removeConnectionFromCommunity(deletion, this.connection.id)
-      }
-
-      this.$emit("update-connection", {
-        ...this.connection,
-        additions,
-        deletions,
-      })
-    },
-    getCommunities(connectionId) {
-      return Object.values(this.communities).filter(community =>
-        community.connections.includes(connectionId)
-      )
-    },
-    getDifferences(original, newVersion) {
-      const additions = newVersion.filter(item => !original.includes(item))
-      const deletions = original.filter(item => !newVersion.includes(item))
-      return { additions, deletions }
+    handlEditConnection(event) {
+      this.handleBack()
+      this.$emit("edit-connection", event)
     },
   },
 }
@@ -247,25 +126,5 @@ export default {
   position: absolute;
   left: 0;
   bottom: 0;
-}
-
-ul {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-
-.form {
-  width: 100%;
-  height: 100%;
-}
-
-.content-wrapper {
-  background: white;
-  position: relative;
-  z-index: 10;
-  height: 100%;
-  border-top: 1px solid var(--cos-color-tertiary);
-  flex-grow: 1;
 }
 </style>
